:poop: extract code into modules and classes
these need to be improved and fixed
