+++ lib/menu.rb
@@ -8,4 +8,5 @@ class Menu
     puts '-----------'
     puts " #{board[6]} | #{board[7]} | #{board[8]}"
   end
+  
 end
